+++ src/canvas/widgets/basic_table_arrows.rs
@@ -54,9 +54,8 @@ impl BasicTableArrows for Painter {
         let left_name = left_table.get_pretty_name();
         let right_name = right_table.get_pretty_name();
 
-        let num_spaces = usize::from(draw_loc.width)
-            .saturating_sub(6 + left_name.len() + right_name.len())
-            as usize;
+        let num_spaces =
+            usize::from(draw_loc.width).saturating_sub(6 + left_name.len() + right_name.len());
 
         let arrow_text = vec![
             Text::Styled(format!("\n◄ {}", left_name).into(), self.colours.text_style),
